+++ driver-core/src/main/java/com/datastax/driver/core/querybuilder/Batch.java
@@ -65,16 +65,17 @@ public class Batch extends BuiltStatement {
      *
      * @param statement the new statement to add.
      * @return this batch.
+     *
+     * @throws IllegalArgumentException if counter and non-counter operations
+     * are mixed.
      */
     public Batch add(Statement statement) {
-        boolean isCounterOp = statement instanceof BuiltStatement
-                && ((BuiltStatement) statement).isCounterOp();
+        boolean isCounterOp = statement instanceof BuiltStatement && ((BuiltStatement) statement).isCounterOp();
 
         if (this.isCounterOp == null)
             setCounterOp(isCounterOp);
         else if (isCounterOp() != isCounterOp)
-            throw new RuntimeException(
-                    "can't mix counter operations and non-counter operations in a batch statement");
+            throw new IllegalArgumentException("Cannot mix counter operations and non-counter operations in a batch statement");
 
         this.statements.add(statement);
         setDirty();