+++ driver-core/CHANGELOG.rst
@@ -3,6 +3,7 @@ CHANGELOG
 
 * 1.0.0-beta2:
     - [JAVA-51] Support blob constants in the query builder.
+    - [JAVA-60] Support counter batches in the query builder.
 
 
 * 1.0.0-beta1: